Mon, Feb 7, 2022, 11:48 PM +00:00

+++ filter-less/helpers.c
@@ -21,10 +21,10 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
 // Convert image to sepia
 void sepia(int height, int width, RGBTRIPLE image[height][width])
 {
-    float average;
-    float sepiaRed;
-    float sepiaGreen;
-    float sepiaBlue;
+    int average;
+    int sepiaRed;
+    int sepiaGreen;
+    int sepiaBlue;
     for (int i = 0; i < height; i++)
     {
         for (int j = 0; j < width; j++)
